split out former index file
