+++ lib/watch.js
@@ -1,109 +1,20 @@
-import fs from 'fs-extra'
 import path from 'path'
 import c from 'ansi-colors'
-import PQueue from 'p-queue'
 import graph from 'watch-dependency-graph'
-import exit from 'exit'
 import globParent from 'glob-parent'
 import chokidar from 'chokidar'
 
-import { debug } from './lib/debug'
-import {
-  CWD,
-  CONFIG_DEFAULT,
-  OUTPUT_DYNAMIC_PAGES_ENTRY
-} from './lib/constants'
-import { createStaticEntry, createDynamicEntry } from './lib/createEntries'
-import { pathnameToHtmlFile } from './lib/pathnameToHtmlFile'
-import { log } from './lib/log'
-import { timer } from './lib/timer'
-import { getFiles, isStatic, isDynamic } from './lib/getFiles'
-import { safeRequire } from './lib/safeRequire'
-import { ignoredFilesArray } from './lib/ignore'
-import { set as setConfig } from './lib/config'
-import { clearMemoryCache } from './load'
+import { debug } from './debug'
+import { CWD, CONFIG_DEFAULT, OUTPUT_DYNAMIC_PAGES_ENTRY } from './constants'
+import { createStaticEntry, createDynamicEntry } from './createEntries'
+import { log } from './log'
+import { getFiles, isStatic, isDynamic } from './getFiles'
+import { safeRequire } from './safeRequire'
+import { ignoredFilesArray } from './ignore'
+import { set as setConfig } from './config'
+import { renderStaticEntries } from './renderStaticEntries'
 
-export async function prepareStaticEntry (entry) {
-  delete require.cache[entry.entryFile]
-
-  const { getPaths, render, createDocument } = require(entry.entryFile)
-
-  return {
-    pages: await getPaths(),
-    render,
-    createDocument
-  }
-}
-
-export function renderStaticEntries (entries, options, done) {
-  return new Promise(async (y, n) => {
-    debug('renderStaticEntries', entries)
-
-    const preparedEntries = []
-    const allGeneratedFiles = []
-    const queue = new PQueue({ concurrency: 10 })
-
-    function onError (e, { location }) {
-      if (options.watch) {
-        log(
-          `\n  ${c.red('error')} ${location}\n\n  > ${e.stack || e}\n\n${c.gray(
-            `  errors detected, pausing...`
-          )}\n`
-        )
-
-        // IMPORTANT, reset this for next pass
-        queue.clear()
-      } else {
-        log(`\n  ${c.red('error')} ${location}\n\n  > ${e.stack || e}\n`)
-      }
-
-      y({ allGeneratedFiles })
-    }
-
-    queue.on('idle', () => {
-      y({ allGeneratedFiles })
-      clearMemoryCache()
-    })
-
-    for (const entry of entries) {
-      try {
-        const p = await prepareStaticEntry(entry)
-        preparedEntries.push(p)
-      } catch (e) {
-        const location = entry.sourceFile.replace(CWD, '')
-        if (e.message.includes(`does not provide an export named 'getPaths'`)) {
-          log(`\n  ${c.yellow('error')} ${location} missing getPaths export\n`)
-        } else {
-          onError(e, { location })
-        }
-      }
-    }
-
-    for (const { pages, render, createDocument } of preparedEntries) {
-      for (const page of pages) {
-        queue.add(async () => {
-          try {
-            const time = timer()
-            const result = await render({ pathname: page })
-            const filename = pathnameToHtmlFile(page)
-
-            allGeneratedFiles.push(filename)
-
-            fs.outputFileSync(
-              path.join(options.output, 'static', filename),
-              createDocument(result),
-              'utf-8'
-            )
-
-            log(`  ${c.gray(time().padEnd(8))}${page}`)
-          } catch (e) {
-            onError(e, { location: page })
-          }
-        })
-      }
-    }
-  })
-}
+import { clearMemoryCache } from '../load'
 
 export async function watch (config, options = {}) {
   debug('watch initialized with config', config)
@@ -341,34 +252,4 @@ export async function watch (config, options = {}) {
    */
   staticWatcher.add(staticIds)
   dynamicWatcher.add(dynamicIds)
-}
-
-export async function build (config, options = {}) {
-  debug('watch initialized with config', config)
-
-  /*
-   * Set computed config for later use
-   */
-  setConfig(config)
-
-  const staticIds = getFiles(config.pages).filter(isStatic)
-  let dynamicIds = getFiles(config.pages).filter(isDynamic)
-
-  if (!staticIds.length && !dynamicIds.length) {
-    log(`  ${c.gray('nothing to build')}\n`)
-    exit()
-  }
-
-  // create initial entries
-  let staticEntries = staticIds.map(id => createStaticEntry(id, config))
-
-  createDynamicEntry(dynamicIds, config)
-
-  options.onRenderStart()
-
-  const { allGeneratedFiles } = await renderStaticEntries(staticEntries, {
-    output: config.output
-  })
-
-  options.onRenderEnd({ count: allGeneratedFiles.length })
 }
